pyproject.toml
[project]
name = "campus-market-place-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "celery>=5.6.2",
    "fastapi[standard]>=0.135.1",
    "pathlib>=1.0.1",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.12.0",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.22",
    "redis>=7.3.0",
    "sqlalchemy>=2.0.48",
]
